Use (u)int_fast32_t when we don't care about the type
This should be 32-bit on 32-bit CPUs and 64-bit on 64-bit CPUs.  It
may be better to use (u)int_fast16_t in case we're running on a
16-bit CPU (e.g., AVR), but I was worried that there may be platforms
where sizeof(int_fast16_t) is smaller than the register size (like
it is for int_fast8_t on my machine), and that would just be a waste.
If you're running SIMDe on a 16 bit CPU please get in touch and we
can revisit the decision.