Remove all C-style casts in x86/x86_64 code.
Use HEDLEY_STATIC_CAST and HEDLEY_REINTERPRET_CAST instead.  In C they
use C-style casts, but in C++ they use static_cast or reinterpret_cast
instead.

Clang has an off-by-default warning, -Wold-style-cast, which is emitted
when a C-style cast is used.  IAR also has a warning for this which is
on at least when --remarks are passed.