mmx: switch to NEON implemantions which don't require immediate mode
GCC was fine with the old version, but clang 8) emits errors.

Unfortunately these could be a bit slower, so I've moved the vector
extension versions to have a higher priority for these functions.
The compiler should be able to generate the faster versions if the
value is known at compile-time, and if not it should generate the
same code as the intrinsics version.