sse: use _mm_move_ss instead of reimplenting repeatedly
There are a lot of functions that follow this pattern; instead of
hardcoding each one, we now just have the native version, then a
fallback on doing the operation on a whole vector, then blending the
first element of the result with the first operand with _mm_move_ss
if a SIMD instruction set extension is found.  Otherwise, just fall
back on the portable version.

This should generate the same code but gets rid of a lot of mostly
repeated implementations, and we can still always implement optimized
versions when available.