Don't use HEDLEY_CONST_CAST in simde_assert_aligned
In C mode HEDLEY_CONST_CAST uses a statement expr so we can disable
the warning about casting away a const qualifier.  This was causing
ICC to think that there was a side-effect (see
https://software.intel.com/en-us/forums/intel-c-compiler/topic/844402
for the bug report) even though there wasn't.

Instead of using HEDLEY_CONST_CAST to cast to uintptr_t, we now just
cast to a const uintptr_t, which is probably a better solution
anyways, and AfAICT all compilers are happy with it.