mmx: include <mmintrin.h> even if MMX is disabled
If we use a later ISA extension, like SSE, compilers will include
<mmintrin.h> automatically to get the __m64 type.  The problem is that
with SIMDE_ENABLE_NATIVE_ALIASES enabled we were creating a typedef
from simde__m64 to __m64, which was incompatible with the native
definition.

This patch will just include <mmintrin.h> on x86, even if MMX has not
enabled.  If native aliases are enabled we should be okay since we just
define macros with the same names as the functions which map to the
SIMDe version, but frankly I do expect a couple of bugs to come out of
this.  Hopefully we can work around them, because without this native
aliases are broken on MSVC on x86_64.

However, there is an exception for Windows on x86 (but not x86_64).
Since merly using the __m64 type in our API can cause incorrect results
we don't do it unless SIMDE_ENABLE_NATIVE_ALIASES is defined (since
with native aliases we can't use our own type as it would break the
native functions.