Allow use without the standard library.
The functionality will be somewhat limited; functions which would
otherwise require the standard library are simply not implemented
and HEDLEY_UNREACHABLE() is used to try to get the compiler to emit
a diagnostic if they are used.

Still, for a lot of applications, especially those that don't require
floating point, this could be useful.