avx512f: reordering and formatting functions
This is going to kill the git blame, but I'd rather do it earlier than
later since it would be even worse in the future.

The goal here is to make it easier to add an entire group of functions
at once and keep closely related functions together to make it a bit
easier to read the code.

I also went through and fixed the indentation of the preprocessor
macros in the functions so they use the new style, but there shouldn't
be any changes in functionality from this patch unless you count test
cases executing in a different order.