More macro cleanups.
This is primarily about using SIMDE_X86_*_NATIVE instead of
SIMDE_ARCH_X86_* as the latter is basically just what we detect from
the compiler wheras the former can be enabled and disabled as needed.

Hopefully this is the last of the changes.  I've been grepping for
different patterns and can't seem to find any more.