common: prefer __alignof__ over alignof
The latter will generate a -Wc++98-compat-pedantic diagnostic in clang,
the underscore version will not.  This happens even though we check to
make sure the feature is supported (via __has_feature or if we're
compiling in C++11+ mode).

This also changes HEDLEY_GCC_VERSION_CHECK to HEDLEY_GNUC_VERSION_CHECK
for detecting this feature.