common: improve isnan/isnanf detection code, especially in C++ mode
In C++ mode, including <cmath> will undefine isnan, but HUGE_VAL will
still be defined.  So if isnan isn't defined but HUGE_VAL is and we're
in C++ mode, we assume std::isnan exists (and does what we want).

This also removes XL C/C++ from the whitelist for __builtin_isnanf
since it implements __builtin_isnan but not __builtin_isnanf.

Fixes #282.