sse2: disable some ARM implementations of _mm_bs{l,r}li_si128 on clang.
Clang doesn't allow you to pass 0 to several functions, even in code
which is never called.  For example:

  (imm8 == 0) ? (a) : vextq_s8(vdupq_n_s8(0), (a), imm8);

Will result in an error on clang if imm8 == 0 even though we handle
the 0 case separately.  All the other compilers allow this.

The workaround is basically to use a non-immediate-mode version.
This won't be quite as fast for some functions, but it's the best
we can do unless clang fixes their compiler.