sse2,avx,avx512f: improve loadu_si* implementations
Clang seems to have lots of problems on ARM.  I think it is assuming
that the argument is aligned due to the type, which is perfectly
reasonable.  This adds a version which goes through a packed, may_alias
struct, which is the method used in clang's internal headers, but
unfortunately I'm still seeing errors on AArch64.  I think that once
you cast to __m128i* (or __256i* or __512i*) you've told clang that
the data is aligned, and there is nothing we can do.  Since those
casts happen in code we can't control thee really isn't anything we
can do.

Unfortunately, poor decisions by Intel make this somewhat difficult.
Using __m{128,256,512}i as the type is completely bonkers.

Basically, if you care about portablility, *do not use*
_mm_loadu_si128 or _mm256_loadu_si256.