neon: work around an ICE in ICC
ICC crashes when preseted with a struct which contains a struct with
an array of vectors.  For details, see
https://software.intel.com/en-us/forums/intel-c-compiler/topic/857088

Unfortunately I can't find a sensible work-around for this, so I'm just
disabling that code on ICC.  Apparently the next version of ICC will be
based on clang, at which point this should work.