sse2: fix mm_andnot_pd on POWER and ARM
This is a pretty serious bug.

The test was looking to see if the results were *close* to the expected
results, not exactly equal.  For most *_pd functions that's desirable,
but in this case the only bits getting flipped were too insignificant
to matter, so the test was basically completely innefective.

That allowed us to add multiple implementations where we took the
complement of b instead of a, meaning the function pretty much did the
opposite of what it was supposed to.