sse2: use NEON _mm_slli_* implementations on clang
We were avoiding them because imm8 may be out of the range expected by
clang. We actually handle those cases separately, but apparently clang
performs these checks before dead code elimination, so it would trigger
errors on clang.  This just adds a useless AND to make clang happy. The
value has to be an ICE anyways, so there is no performance penalty.