sse4.1: add SSE2 and shuffle-based fallbacks for _mm_cvtepi*_epi*
For unsigned conversions, we just shuffle the inputs into the right
positions, and all other elements are set to zero.  For signed
conversions we first shuffle the inputs into the most significant
portion of each output lane, then use an arithmetic shift to move
them to the least significant section.

Obviously this all depends on being endianness.  I implemented these
for little endian only, but it should be possible to tweak them a bit
to get a good big endian implementation.

The good news is that this seems to be a *lot* faster than even the
__builtin_convertvector implementations, and works well even on SSE2.
The biggest problem is that on x86 there is no 64-bit right shift
until AVX-512, at which point we would just be using the native
implementations anyways, so conversions to 64-bit signed integers
are still slow on x86.  Luckily most other architecutres support
64-bit shifts.