sse: add SIMDE_MM_{GET,SET}_FLUSH_ZERO_MODE
This also adds a bunch of related native aliases which were missing.

There are no tests for this because it's not something we can feasibly
emulate on platforms which don't support it, and few platforms support
it.

It *may* be possible to implement this on AArch64.  It is controlled
by the FPCR register, but ACLE documentation says to access the FPCR
register using fegetenv/fesetenv, and AFAICT glibc doesn't expose a
way to access the relevant bits.  I think you can get at it via the
fpsimd_context struct in sigcontext.h, but portability is likely to
be a problem.  IMHO it is better to just leave things alone.

AFAIK there is no way to do this on 32-bit Arm.  I haven't looked into
other architectures, and given the limited utility I don't plan to.  I
am happy to accept patches if someone figures out a good way to do it
on other architectures.