avx512/insert: convert macros to functions, regenerate old-style tests
clang 12 on PPC64LE was generating bad code for the old version.  The
first test function clobbered the iterator variable of the loop to
call each test function (in the footer), which eventually caused a
segfault.  Of course it didn't happen reliably, but generally at
least every few runs (at -O2, never at -O0).

Getting a good reduced case was tricky, and I can't really understand
PPC assembly, so the best I could come up with was migrating the tests
to the new style with arrays of elements instead of calling set funcs
in the initializer.  However, the CONSTIFY macros don't work with
macros, so I had to switch the insert implementations from macros to
functions.

Fixes #741.