Initial import of a portable SVE implementation.
This is just a starting point, nothing really useful yet.  Currently
there is basically only support for add and cmplt, but the main thing
is that this sets up all the infrastructure.  Overall, I think this is
working out pretty well, though I do see a few potential problem spots.

There are basically three implementations; native SVE, AVX-512, and
everything else.  The everything else implementation is pretty
straightforward; the biggest drawback there is likely going to be the
lack of support for predicated operations, which means that in addition
to all the bitwise AND ops (basically one for every operation) there
will also be a bit of extra pressure on the SIMD registers since one
is likely to be dedicated to the predicate.  Bitwise AND is pretty
cheap, so I think this should perform pretty well.

The AVX-512 implementations are interesting because there is a weird
inversion where AVX-512 has multiple types for the predicate
(__mmask8/16/32/64), but SVE only has a single type (svbool_t).
We store everything in an __mmask64, but *really* converting is not
trivial.  Ideally we can use pdep/pext, but that requires BMI2 and
AVX-512 doesn't imply BMI2 (though I *think* it is pretty safe to
assume).  Still, it's relatively expensive.  Luckily, I think most
compilers will be able to elide most of the conversions, at least
unless people use svbool_t in APIs where the compiler is unable to
reason about code on the other side.

Native SVE might be tricky.  Obviously, when we can just pass through
to native calls there isn't an issue, but when implementing SVE2 using
SVE I think there are going to be a lot of move instructions. Composing
SVE2 operations using SVE shouldn't be bad, but if we have to switch to
NEON or scalar and iterate through the vector we'll have to load and
store each input (including the predicate) several times for each
function.

One more thing worth mentioning is that I really don't like svptest; it
would be much better to have an unpredicated version of the function
since most calls are likely to just use svptrue for the predicate, and
the predicate adds a lot of complexity to the implementation.  The good
news is that the compiler can generally get get rid of a lot of the
overhead, but I'd still feel a lot better if we didn't have to depend on
that.  I may add some additional functions as extensions which omit the
predicate, which could be a significant win for non-SVE targets.

Considering how different SVE is from the other ISA exxtensions we deal
with I'm pleasantly surprised by how few real problems we are likely to
run into are.  Implementing more functionality should be relatively
straightforward, and this would be an excellent way for people to take
advantage of AVX-512 and AVX2 (and SVE, NEON, AltiVec, etc.).

Note that, at compile time, you can choose to use 256-bit vectors with
AVX-512VL instead of 512-bit vectors by defining the
SIMDE_NATURAL_VECTOR_SIZE macro to 256 prior to including SIMDe.  This
can help avoid Intel throttling back the CPU and can (somewhat counter-
intuitively) increase performance.  This isn't the same thing as just
disabling AVX-512 and using AVX2 since we can still make use of the
predicated instructions (e.g., _mm256_mask_add_ps instead of an
_mm256_add_ps and an _mm256_and_ps).