Fix compilation with clang on POWER
This squashes a few warnings, but the main thing is that there is a bug
with the 64-bit integer version of vec_sel in clang prior to version 12
which causes an error.  The good news is that it's fixed, so now we can
add an end version for that bug macro.