avx512/range: remove CONSTIFY macro usage
The major problem here was that the calls to shorter versions in the
callbacks didn't work outside of the tests (where we disable the
SIMDE_REQUIRE_CONSTANT macros so we can get beter coverage); for
example, even including avx512.h would result in compile-time errors.

At first I tried to just use constify for those cases, plus a few
tricks to reduce duplication, but it was still to much for some of our
CI builds (specifically, the static analysis builds), so I had to move
the fallbacks into macros outside of the function definiion.  That also
meant I had to change the tests to get rid of CONSTIFY there since it
would no longer work on MSVC.

The implementations part of this looks worse than it is because the
indentation had to change when I removed the fallbacks; the code didn't
really change all that much.