avx512/set, avx512/popcnt: use _mm512_set_epi8 only when available
The IIG says it's an AVX-512F function (not AVX-512BW), but it was
only added to clang in version 5, and GCC in version 10.  This changes
the new popcnt implementation to use the simde-prefixed versiion, and
changes the implementation of simde_mm512_set_epi8 to actually use
_mm512_set_epi8, but only on clang >= 5 or gcc >= 10.  We can probably
add some other compilers in the future, but TBH I don't think it really
matters since the generated code will most likely be the same.