sse, sse2: clean up several shuffle macros
There were several NEON fallbacks which depended on statement exprs
but didn't properly test for them, which means they were broken on
MSVC.  Furthermore, they used input parameters (other than the imm8
parameter), which could lead to unintentionally evaluating the parameter
multiple times, and they declared variables (like "ret") which were
likely to shadow variables in the code they were used in.

This change properly tests for statement exprs and, if they are not
supported, falls back on the portable functions.  The good news is that
this really only effects MSVC since virtually all other compilers
support statement exprs.  The bad news is that it effects MSVC.
Unfortunately I don't think there is a way to do this in MSVC.