avx2: separate natural vector length for float, int, and double types
Since Intel had a gap of a couple of years where they supported 256-bit
vectors of floats but not integer types, in some cases we want to fall
back to 128-bit vectors even though 256-bit vectors are actually
available.

Unfortunately this is likely to be a fairly significant performance hit
since it will involve a lot of shuffling the high half into a different
register, performing the 128-bit operation, then shuffling it back.  My
instinct is that it would actually be faster for projects which are
hitting this path heavily (basically, -mavx but not -mavx2, calling
lots of AVX2 functions) to just disable AVX altogether.  Still, this
could be useful if there are just a few AVX2 calls mixed in with mostly
AVX.